// src/app/core/utils/Interfaces/forms/existing-coverage.interface.ts
export interface ExistingCoverages {
  B1_ExistingLifeInsurance: string | undefined;
  B1_DisabilityInsuranceInPercentage: string | undefined;
  B1_ExistingCriticalIllnessInsurance: string | undefined;

  B2_ExistingLifeInsurance: string | undefined;
  B2_DisabilityInsuranceInPercentage: string | undefined;
  B2_ExistingCriticalIllnessInsurance: string | undefined;
}