// src/app/core/utils/Interfaces/forms/monthly-expenses.interface.ts
export interface MonthlyExpenses {
  NewMonthlyMortgageLoanPayment: string | undefined;
  ExistingMonthlyLiabilityDebtPayments: string | undefined;
  OtherMonthlyExpensesInPercentage: string | undefined;
  OtherMonthlyExpensesInNumber: string | undefined;

  TotalMonthlyIncome: string | undefined;
  TotalMonthlyExpenses: string | undefined;
  NetMonthlyIncome: string | undefined;
}